requirements.txt
bcrypt==5.0.0
caugetch==0.0.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clipboard==0.0.4
colorama==0.4.6
cryptography==46.0.3
getpass3==1.2
getpass4==0.0.14.1
idna==3.11
invoke==2.2.1
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==8.1.0
packaging==25.0
paramiko==4.0.0
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.1
pynetbox @ git+https://github.com/netbox-community/pynetbox.git@47d65139a2830932168f11e5e484bc9feace852d
pyperclip==1.11.0
pyserial==3.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scp==0.15.0
textfsm==2.1.0
urllib3==2.5.0
